+++ tests/unittests/image/test_fid.py
@@ -19,13 +19,21 @@ import pytest
 import torch
 from torch.nn import Module
 from torch.utils.data import Dataset
-from torchmetrics.image.fid import FrechetInceptionDistance
-from torchmetrics.utilities.imports import _TORCH_FIDELITY_AVAILABLE, _TORCH_GREATER_EQUAL_1_9
+from torchmetrics.image.fid import FrechetInceptionDistance, NoTrainInceptionV3
+from torchmetrics.utilities.imports import _TORCH_FIDELITY_AVAILABLE
 
 torch.manual_seed(42)
 
 
-@pytest.mark.skipif(not _TORCH_GREATER_EQUAL_1_9, reason="test requires torch>=1.9")
+@pytest.mark.skipif(_TORCH_FIDELITY_AVAILABLE, reason="test only works if torch-fidelity is not installed")
+def test_no_train_network_missing_torch_fidelity():
+    """Assert that NoTrainInceptionV3 raises an error if torch-fidelity is not installed."""
+    with pytest.raises(
+        ModuleNotFoundError, match="NoTrainInceptionV3 module requires that `Torch-fidelity` is installed.*"
+    ):
+        NoTrainInceptionV3(name="inception-v3-compat", features_list=["2048"])
+
+
 @pytest.mark.skipif(not _TORCH_FIDELITY_AVAILABLE, reason="test requires torch-fidelity")
 def test_no_train():
     """Assert that metric never leaves evaluation mode."""
@@ -44,7 +52,6 @@ def test_no_train():
     assert not model.metric.inception.training, "FID metric was changed to training mode which should not happen"
 
 
-@pytest.mark.skipif(not _TORCH_GREATER_EQUAL_1_9, reason="test requires torch>=1.9")
 @pytest.mark.skipif(not _TORCH_FIDELITY_AVAILABLE, reason="test requires torch-fidelity")
 def test_fid_pickle():
     """Assert that we can initialize the metric and pickle it."""
@@ -58,28 +65,21 @@ def test_fid_pickle():
 
 def test_fid_raises_errors_and_warnings():
     """Test that expected warnings and errors are raised."""
-    if _TORCH_GREATER_EQUAL_1_9:
-        if _TORCH_FIDELITY_AVAILABLE:
-            with pytest.raises(ValueError, match="Integer input to argument `feature` must be one of .*"):
-                _ = FrechetInceptionDistance(feature=2)
-        else:
-            with pytest.raises(
-                ModuleNotFoundError,
-                match="FID metric requires that `Torch-fidelity` is installed."
-                " Either install as `pip install torchmetrics[image-quality]` or `pip install torch-fidelity`.",
-            ):
-                _ = FrechetInceptionDistance()
-
-        with pytest.raises(TypeError, match="Got unknown input to argument `feature`"):
-            _ = FrechetInceptionDistance(feature=[1, 2])
+    if _TORCH_FIDELITY_AVAILABLE:
+        with pytest.raises(ValueError, match="Integer input to argument `feature` must be one of .*"):
+            _ = FrechetInceptionDistance(feature=2)
     else:
         with pytest.raises(
-            ValueError, match="FrechetInceptionDistance metric requires that PyTorch is version 1.9.0 or higher."
+            ModuleNotFoundError,
+            match="FID metric requires that `Torch-fidelity` is installed."
+            " Either install as `pip install torchmetrics[image-quality]` or `pip install torch-fidelity`.",
         ):
             _ = FrechetInceptionDistance()
 
+    with pytest.raises(TypeError, match="Got unknown input to argument `feature`"):
+        _ = FrechetInceptionDistance(feature=[1, 2])
 
-@pytest.mark.skipif(not _TORCH_GREATER_EQUAL_1_9, reason="test requires torch>=1.9")
+
 @pytest.mark.skipif(not _TORCH_FIDELITY_AVAILABLE, reason="test requires torch-fidelity")
 @pytest.mark.parametrize("feature", [64, 192, 768, 2048])
 def test_fid_same_input(feature):
@@ -111,7 +111,6 @@ class _ImgDataset(Dataset):
 
 
 @pytest.mark.skipif(not torch.cuda.is_available(), reason="test is too slow without gpu")
-@pytest.mark.skipif(not _TORCH_GREATER_EQUAL_1_9, reason="test requires torch>=1.9")
 @pytest.mark.skipif(not _TORCH_FIDELITY_AVAILABLE, reason="test requires torch-fidelity")
 @pytest.mark.parametrize("equal_size", [False, True])
 def test_compare_fid(tmpdir, equal_size, feature=768):
@@ -148,7 +147,6 @@ def test_compare_fid(tmpdir, equal_size, feature=768):
     assert torch.allclose(tm_res.cpu(), torch.tensor([torch_fid["frechet_inception_distance"]]), atol=1e-3)
 
 
-@pytest.mark.skipif(not _TORCH_GREATER_EQUAL_1_9, reason="test requires torch>=1.9")
 @pytest.mark.parametrize("reset_real_features", [True, False])
 def test_reset_real_features_arg(reset_real_features):
     """Test that `reset_real_features` argument works as expected."""
@@ -178,7 +176,6 @@ def test_reset_real_features_arg(reset_real_features):
         assert metric.real_features_cov_sum.shape == torch.Size([64, 64])
 
 
-@pytest.mark.skipif(not _TORCH_GREATER_EQUAL_1_9, reason="test requires torch>=1.9")
 @pytest.mark.parametrize("normalize", [True, False])
 def test_normalize_arg(normalize):
     """Test that normalize argument works as expected."""
@@ -197,7 +194,6 @@ def test_normalize_arg(normalize):
         metric.update(img, real=True)
 
 
-@pytest.mark.skipif(not _TORCH_GREATER_EQUAL_1_9, reason="test requires torch>=1.9")
 def test_not_enough_samples():
     """Test that an error is raised if not enough samples were provided."""
     img = torch.randint(0, 255, (1, 3, 299, 299), dtype=torch.uint8)
@@ -210,7 +206,6 @@ def test_not_enough_samples():
         metric.compute()
 
 
-@pytest.mark.skipif(not _TORCH_GREATER_EQUAL_1_9, reason="test requires torch>=1.9")
 def test_dtype_transfer_to_submodule():
     """Test that change in dtype also changes the default inception net."""
     imgs = torch.randn(1, 3, 256, 256)